requirements.txt
# Не ставить пакет `serial` с PyPI — это не pyserial и ломает `serial.Serial`.
pyserial>=3.5
pytest>=7.0
